Update netlify links to github.io
It seems to me that the netlify website is not anymore and I see that the official url next to the Description has been changed too. We should update netlify references from Readme to point to the new location.

I had the netlify URL bookmarked and I did not notice that there are new Icons already and also the web is different.